Release 1.2.0 with GTP-U and PPPoE support (needs to be enabled in P4…

+++ loadGenerators/iperf3/scripts/install_iperf3_at_loadgen.sh
@@ -13,18 +13,24 @@ check_iperf_version() {
 
 result=$(check_iperf_version)
 if [ $result = "ok" ]; then
-  printf "IPerf3 is already installed!"
+  printf "IPerf3.1.3 or newer is already installed!"
 else
-
-  printf "\n-----------------------------------------\n"
-  printf "Installing iPerf3"
-  sudo apt update
-  sudo apt remove -y iperf3 libiperf0
-  wget https://iperf.fr/download/ubuntu/libiperf0_3.1.3-1_amd64.deb
-  wget https://iperf.fr/download/ubuntu/iperf3_3.1.3-1_amd64.deb
-  sudo dpkg -i libiperf0_3.1.3-1_amd64.deb iperf3_3.1.3-1_amd64.deb
-  rm libiperf0_3.1.3-1_amd64.deb iperf3_3.1.3-1_amd64.deb
-  printf "\n-----------------------------------------\n"
+  . /etc/lsb-release
+  if [ $DISTRIB_RELEASE = "16.04" ]; then
+    printf "\n-----------------------------------------\n"
+    printf "Installing iPerf3 from sources"
+    sudo apt update
+    sudo apt remove -y iperf3 libiperf0
+    wget https://iperf.fr/download/ubuntu/libiperf0_3.1.3-1_amd64.deb
+    wget https://iperf.fr/download/ubuntu/iperf3_3.1.3-1_amd64.deb
+    sudo dpkg -i libiperf0_3.1.3-1_amd64.deb iperf3_3.1.3-1_amd64.deb
+    rm libiperf0_3.1.3-1_amd64.deb iperf3_3.1.3-1_amd64.deb
+    printf "\n-----------------------------------------\n"
+  else
+    printf "\n-----------------------------------------\n"
+    printf "Installing iPerf3 from apt"
+    sudo apt -y install iperf3;
+  fi
 fi
 printf "Installing ethtool"
 sudo apt -y install ethtool iproute2 net-tools lsof
